requirements.txt
djangorestframework==3.16.0
drf-yasg==1.21.10
psycopg2-binary~=2.9.10
